borg-assimilate: ask user what to do with existing gitdir
If the gitdir already exists but the working tree does not, then offer
to either reuse it as-is or to replace it with a new clone.  Only the
latter option guarantees that the configuration inside the submodule
is still up-to-date.

Previously the user had to use a prefix argument to reuse the existing
gitdir.  And it was not possible to replace it.  The user had to first
manually deleting the existing gitdir.

When the worktree exists in addition to the gitdir, then just call git
without any preparations, which causes the existing gitdir to be used
as-is.  We don't offer other choices because this usually happens
after `borg-clone' was used to preview the package before assimilation
and a prompt would be annoying in this case.