borg--maybe-absorb-gitdir: use "git submodule absorbgitdirs"
The old implementation assumed that the super repository's own gitdir
is located inside its working tree.  But that doesn't have to be true.
In my case for example it is located at "~/.git/modules/emacs.d/" and
the gitdirs of packages are at "~/.git/modules/emacs.d/modules/<pkg>".

The "absorbgitdirs" subcommand was added in v2.12.0.  For older Git
versions keep using the old emacs-lisp implementation.  That is
exactly the kind of thing I don't want to do in this code base.  There
is a correct way of doing this but for compatibility reasons I have to
keep an inferior implementation around.  Please update -- this will be
removed eventually.