Add support for external handlers for build steps.
Introduce a new variable borg-run-build-steps-function to select a
function to run submodules.DRONE.build-step, a default implementation
which runs each command through shell-command and a helper function
borg-build-step which takes a string and a function and pass this
string to this function unless the string starts with a "(", in which
case it (eval) it instead.

This doesn't change the actual behavior of Borg in any way.