borg-clone: return only directories
The new implementation assumes that every directory whose name does
not begin with a dot is a clone.  It does not make sure that it is
an actual repository since a directory that is not a repository yet
might just be an extremely young package.

The old implementation tried to filter out non-repositories, but
failed to do so.