Add files borg.mk and borg.sh
These files used to be part of the Emacs.g collective as "Makefile"
and "bin/borg-bootstrap" and all configurations that did not derive
from that collective had to import those files and keep them manually
up-to-date.

The old "Makefile" of the Emacs.g collective is replaced with one
that sources the "borg.mk" from here.  It also defines one simple
target `bootstrap-borg' itself, which is required to make "borg.mk"
and "borg.sh" available during bootstrap.  That target is fairly
simple and hopefully won't have to be updated often, but independent
configurations still have integrate a copy of this file, because we
cannot bootstrap out of thin air and `git submodule' isn't up to the
task yet.