make: Add kludge for single drone targets
Because the `lib/%' target is now defined in a file located in
a sub-directory, `make' started to additionally build a target
`lib/borg.mk'.  Replace that target with a new `lib/%/build'.

Hopefully we can find a better kludge so that we can go back to
`make lib/MODULE' instead of the current `make lib/MODULE/build'.