borg.mk: Remove obsolete byte-compiler appeasements
Some Emacs development versions declared `if-let' and `when-let'
obsolete.  However it is not possible to use the replacements
`if-let*' and `when-let*' instead for many years to come because
older Emacs versions don't have those.

That resulted in a gazillion of warnings about the former pair being
obsolete, which had to be suppressed because it drowned noteworthy
warnings in a see of noise.

Later the Emacs developers were convinced that this was an issue and
reverted the deprecation.  So we can remove out kludge now.

Re https://github.com/emacscollective/epkg/pull/25