borg-initialize: Keep message about init.el being compiled
When `auto-compile-on-load-mode' compiles "init.el", then we want
to display the message about that for the whole duration because
it takes a while.

Unfortunately `inhibit-message' does not just prevent a new message
from being displayed; it also clears the current message from the
echo area.  That is very much not what we want, so we have to jump
through some additional hoops.