borg--build-noninteractive: Load etc/borg/config.el here as well
Previously this file was only loaded in `borg--build-interactive',
which is strange because in this case one could just as well put
Borg setting in the user-init-file for example (except that the
Borg specific file is reloaded every time this function is called,
which also seems unnecessary in this case).

`borg--build-noninteractive' on the other hand does not have access
to regular user settings, so here it actually makes sense to load
an additional file with Borg specific settings.

For backward compatibility `borg--build-interactive' continues to
load the Borg specific file.  If the file contains settings that
should not apply in both cases, then `noninteractive' can be used
as a condition.