Handle autoloads and load-path more robustly
We cannot rely on autoload files adding entries to `load-path'.  Until
Emacs 26.1 [1: d6d5020c25] `autoload-rubric' did not do so.  For that
reason Borg did not use that function until now and reimplemented its
functionality from scratch.  A consequence was that our version did
not benefit from upstream improvements.

More importantly though even if Borg ensures that entries are added to
`load-path' that does not guarantee that autoload files contain them.
If they are created using the package's own build tools, because
either the user or Borg uses that, then all bets are off.  Org's
Makefile, for example, still doesn't do it.  Also I am currently
changing the Makefiles of all my packages to use `autoload-rubric'
(for Borg in this commit).  When using Emacs 25 then the autoload
files generated using these Makefiles won't extend the `load-path'
anymore either.

Borg now explicitly extends the `load-path' in `borg-activate'.  When
using a recent Emacs release, then that usually is a no-op.  Luckily
this duplicated work it is fairly cheap because new entries are added
to the front of `load-path'.

1: 2017-05-09 d6d5020c2593a1e8ac2fe7ef4f217cfbcacfd32d
   Don't duplicate autoload code in package.el