Remove note about Borg not actually depending on Emacs 26
The only reason why Emacs 25 is still supported is that some of my
other packages still claim to support it (because users complained
when I tried to remove support for old releases) and I occasionally
need to be able to verify whether that is actually still true.  The
only reason for someone to use Borg with Emacs 25 is to do the same
for the packages they maintain.  For all other purposes this package
*does* depend on Emacs 26.

After spending two weeks or more dealing with backward compatibility,
I must say that I would really appreciate it if everybody just updated
their Emacs in a timely fashion, instead of not doing that because
"it's too much work", while at the same time expecting to be able to
conveniently use the latest versions of my packages.  The work cannot
be avoided.  It can only transformed and offloaded to the maintainer,
who might get a bit grumpy as a result.  Not that Borg users fall into
this category, but I had to write it down somewhere.