borg-update-autoloads: Silence a warning for ht package
I don't know how to fix it and even if I did, I would still have to
deal with the fact that the maintain is quite unresponsive.  I have
seen the warning many times and don't want to see it again.

Here it is in it's full glory, just in case I ever want to see it
again:

  ht:0: Warning: Not registering prefix "ht".  Affects: ("ht"
  "ht<-alist" "ht<-plist" "ht?" "ht-get" "ht-get*" "ht-update"
  "ht-update!" "ht-merge" "ht-map" "ht-keys" "ht-values" "ht-items"
  "ht-amap" "ht-aeach" "ht->plist" "ht->alist" "ht-to-plist"
  "ht-to-alist" "ht-p" "ht-create" "ht-clear" "ht-clear!" "ht-copy"
  "ht-contains?" "ht-contains-p" "ht-size" "ht-set" "ht-set!"
  "ht-select" "ht-select-keys" "ht-remove" "ht-remove!" "ht-reject"
  "ht-reject!" "ht-delete-if" "ht-from-alist" "ht-from-plist"
  "ht-find" "ht-each" "ht-empty?" "ht-empty-p" "ht-equal?"
  "ht-equal-p")