Remove support for borg.collective
There were two modes, one for maintainers of a collective and one for
users.  As the maintainer of the "Emacsg" collective, I am not using
the maintainer mode; and I am not aware of any other collectives.
Using the user mode with the "Emacsg" collective isn't useful either,
because I don't push any patched packages for that.