borg-update-autoloads: Add to load-path again
Autoload files are supposed to add the containing directory to the
`load-path'.  This was lost when `autoload' was deprecated in favor
of `loaddefs-gen'.  This is a known issue (see bug#63625).  Use the
same workaround as used by `package-generate-autoloads'.  Instead of
relying on `loaddefs-generate' to do it on its own, pass appropriate
EXTRA-DATA to it.

Closes #159.