Soft-require comp
Before performing native compilation we use `fboundp' to check the
availability of certain native-compilation functions.  This served
two purposes; to suppress warning from the byte-compiler, when
these functions are not available, and to determine whether
native-compilation is available.

In Emacs 30 the latter did not work reliably anymore as the libraries
in which these functions are defined are no longer loaded up front.
We address this by loading `comp' up front when `borg' is loaded,
effectively undoing this upstream optimization.

I've considered to delay loading `comp' until Borg is about to compile
a package, but when native-compilation is not available we would end up
(unsuccessfully) searching for `comp' every time a file is compiled,
which seems more wasteful than potentially needlessly loading `comp'.

(Going forward we might be able to load only `comp-run', but because
`comp--native-compile' is defined in `comp', we cannot do that just
yet.  See `comp-run's library commentary.)

Closes #161.