borg--build-interactive: Use display-buffer
Do this so that users can customize the behavior to their liking.

Previously the buffer was displayed in the selected window, now it is
usually displayed in another window, possibly created by splitting an
existing window.

As a result the likelihood decreases that the whole output is visible,
but we keep this change in behavior because it matches the behavior of
similar functionality in Emacs.

Do however take care to select the window used to display the buffer,
so that it can still be quickly dismissed by merely pressing "q".

Users who want the old behavior back, can do so using something like:

  (setq display-buffer-alist
        (cons '("\\*Borg Build\\*" (display-buffer-same-window))
              display-buffer-alist))

Closes #160.