borg.sh: Drop unsupported and ineffective argument to realpath
The `realpath' on macOS (and likely elsewhere) lacks this argument.

We wanted to use a relative path so that `git clone' would store that
relative path in "<module>/.git"; turns out it converts any relative
path into an absolute path.

In other words, we lose nothing by dropping this argument.

Closes #167.