borg.sh: Improve checking if module is active
When "submodule.<name>.active" is unset, Git considers <name> active.
Do the same here.

Before that variable was added Git, instead checked whether
"submodule.<name>.url" was non-empty.  In fact it still does that also,
so do that here to.

However, do not also consider "submodule.active" and also do not check
if the path to module is intact.  See "is_tree_submodule_active()" in
submodule.c and all the helpers it uses, to learn what it would take
for full compliance.  It's not worth it.

We split the "git submodule update" step into two sub-steps "clone" and
"checkout".  The second step should not check the above variable again.
And the first step should not communicate failure to the second step by
setting "submodule.<name>.active"; only the user should ever set that to
"false".

Instead check whether "<path>/.git" exists.  If the first step succeeded
then it does, otherwise it doesn't.  We merely check if the path exists;
"is_submodule_populated_gently()" from submodule.c, checks its validity
more thoroughly, but that's not worth the effort.