borg-batch-rebuild: Skip empty drones
When cloning a drone fails during bootstrap, we end up with an empty
directory.  If we proceed to build that drone anyway, we end up with
a directory that is no longer empty (it contains at least the autoload
file).

If we then clone again, for example after updating the URL, that fails
because "fatal: destination path 'lib/compat' already exists and is not
an empty directory."  Prevent by not attempting to build empty drones.