make: Add redo target and slightly increase verbosity
The new "redo" target (which depends on "clean" and "build") is similar
to "quick" (which depends on "quick-clean" and "quick-build"), but
builds all drones.  (When I added "quick" I believed "redo" was already
available, but it turned out I had only added that to my own setup.)

Start printing messages at the beginning of all four sub-targets and
spacers at the end of the clean targets, because they all individually
load `borg' and call `borg-initialize'.  This results in duplicated
messages, which one might interpret as an indication of some issue.
The addition of these messages makes that mistaken impression less
likely.