borg.el: Undeprecate if-let and when-let
Deprecating these forms is wrong.  I will keep using them.

In bigger packages I am switching to the implementations provided
by my `cond-let' package.  For smaller packages I am not doing the
same -- at least not yet.  I won't make `borg' depend on `cond-let'
because that should not have any dependencies whatsoever.

Previously users had to live with these warnings or figure out how
to undeprecate them.  Now we take care of that for them.

One might argue that we shouldn't do that because then some package
maintainers won't become aware that they are using obsolete forms.
However, I consider it a good thing if an appropriate `if-let' is not
replaced with an inappropriate `if-let*', so I consider that a win.